// src/types/entities/profile/enums.ts
/**
 * @file Enums para la entidad Profile
 * @module types/entities/profile/enums
 */

/**
 * Enum para el tema del perfil
 */
export enum ThemeMode {
	SYSTEM = 'system',
	LIGHT = 'light',
	DARK = 'dark',
}

/**
 * Enum para el idioma del perfil
 */
export enum Language {
	SPANISH = 'es',
	ENGLISH = 'en',
	PORTUGUESE = 'pt',
	FRENCH = 'fr',
}